Feature section lock with one dummy record

See merge request protocol-dig/pd-ui-backend!349

+++ app/qc_ingest/model/iqvsectionlock_db.py
@@ -1,10 +1,14 @@
-from sqlalchemy import Column, DateTime
+from sqlalchemy import Column, DateTime, and_
 from datetime import datetime
 from .__base__ import SchemaBase, schema_to_dict, MissingParamException
 from ...crud.pd_user import CRUDUserSearch
 from ...models.pd_user import User
 from sqlalchemy.dialects.postgresql import TEXT
+from sqlalchemy.exc import IntegrityError
+import logging
+from app.utilities.config import settings
 
+logger = logging.getLogger(settings.LOGGER_NAME)
 
 class IqvsectionlockDb(SchemaBase):
     __tablename__ = "iqvsectionlock_db"
@@ -45,12 +49,12 @@ class IqvsectionlockDb(SchemaBase):
         """
         if not data.get('link_id', None) and not data.get('doc_id', None):
             raise MissingParamException(f'link_id or doc_id ')
-        
+        doc_id = data.get('doc_id')
         if data.get('section_lock') == False:
             section_info = IqvsectionlockDb()
             crud_user_search = CRUDUserSearch(User)
             section_info.link_id = data.get('link_id')
-            section_info.doc_id = data.get('doc_id','')
+            section_info.doc_id = doc_id
             section_info.userId = user_id = data.get('userId','')
             user_name_obj = crud_user_search.get_by_username(session, user_id)
             if not user_name_obj:
@@ -63,10 +67,21 @@ class IqvsectionlockDb(SchemaBase):
             obj = session.query(IqvsectionlockDb).filter(
                 IqvsectionlockDb.link_id == data['link_id']).first()
             if not obj:
-                raise MissingParamException("{0} in iqv section lock DB".format(data['link_id']))
-            obj.link_id = data['link_id'] = ""
-            obj.last_updated = data['last_updated'] = datetime.utcnow() 
-            session.delete(obj)
+                raise MissingParamException("{0} in iqv section lock DB".format(data['link_id'])) 
+            
+            doc_obj = session.query(IqvsectionlockDb).filter(and_(
+                IqvsectionlockDb.doc_id == doc_id, IqvsectionlockDb.link_id == '', IqvsectionlockDb.userId == '')).first()
+            if not doc_obj: 
+                obj.link_id = data['link_id'] = ''
+                obj.userId = data['userId'] = ''
+                obj.user_name = data['user_name'] = ''
+                obj.last_updated = data['last_updated']  = datetime.utcnow()
+                try:
+                    session.add(obj)
+                except IntegrityError as ex:
+                    logger.exception(f'while adding new doc record got exception : {ex}')
+            else:
+                session.delete(obj)
         return data
 
     @staticmethod
